+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 0.9.0 - 2023-04-25
+- Add CLI for Persephone as two commands that can be made using `make bundle`; command options can be viewed using `--help`.
+  - `validate`: Performs Statement Template validation for a single Statement
+  - `match`: Performs Pattern matching for a Statement batch
+- Move Clojure and ClojureScript dependencies out of main deps into alias (extra) deps
+- Added printing for pattern match errors (not just failures)
+- Change Statement validation error message header and assert message from `Invalid Statement` to `Statement Validation Failure`
+- Fix bug where the `:selected-patterns` keyword arg for `compile-profiles->fsms` did not work
+- Add `:validate-not-empty?` kwargs to compilation functions that are `true` by default; these will result in `::assert/no-templates` or `::assert/no-patterns` exceptions being thrown if no Template validators or Pattern FSMs, respectively, are present after compilation.
+
 ## 0.8.4 - 2023-04-19
 - Update README and API function docstrings.